some restructuring, swagger tick
