+++ .gitignore
@@ -1,1 +1,7 @@
+*.suo
+*.user
+_ReSharper.*
+bin
+obj
+packages
 Dockerfile
