+++ CollibraTests/UnitTest1.cs
@@ -1,0 +1,18 @@
+using NUnit.Framework;
+
+namespace CollibraTests
+{
+    public class Tests
+    {
+        [SetUp]
+        public void Setup()
+        {
+        }
+
+        [Test]
+        public void Test1()
+        {
+            Assert.Pass();
+        }
+    }
+}
