+++ HelloRadix.sln
@@ -13,6 +13,12 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HelloRadixTests", "HelloRadixTests\HelloRadixTests.csproj", "{4DF40E64-51D7-45FC-A99D-B6DC79478CC2}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Common", "Common\Common.csproj", "{950C8670-04B9-422D-8ADC-B877A2389978}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Collibra", "Collibra\Collibra.csproj", "{B4999FD9-089D-498B-8170-1AE4CA3AAFFF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CollibraTests", "CollibraTests\CollibraTests.csproj", "{D4B8705A-9BF9-4BCC-B1AE-CEF3ED6216E0}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +33,18 @@ Global
 		{4DF40E64-51D7-45FC-A99D-B6DC79478CC2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4DF40E64-51D7-45FC-A99D-B6DC79478CC2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4DF40E64-51D7-45FC-A99D-B6DC79478CC2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{950C8670-04B9-422D-8ADC-B877A2389978}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{950C8670-04B9-422D-8ADC-B877A2389978}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{950C8670-04B9-422D-8ADC-B877A2389978}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{950C8670-04B9-422D-8ADC-B877A2389978}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B4999FD9-089D-498B-8170-1AE4CA3AAFFF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B4999FD9-089D-498B-8170-1AE4CA3AAFFF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B4999FD9-089D-498B-8170-1AE4CA3AAFFF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B4999FD9-089D-498B-8170-1AE4CA3AAFFF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D4B8705A-9BF9-4BCC-B1AE-CEF3ED6216E0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D4B8705A-9BF9-4BCC-B1AE-CEF3ED6216E0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D4B8705A-9BF9-4BCC-B1AE-CEF3ED6216E0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D4B8705A-9BF9-4BCC-B1AE-CEF3ED6216E0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
