+++ HelloRadix/Controllers/CollibraController.cs
@@ -4,7 +4,8 @@ using System;
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
-using HelloRadix.Exceptions;
+using Collibra.Services;
+using Common.Exceptions;
 using Microsoft.AspNetCore.Authorization;
 
 namespace HelloRadix.Controllers
