--- HelloRadix/bin/Debug/netcoreapp3.1/HelloRadix.StaticWebAssets.xml
@@ -1,2 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<StaticWebAssets Version="1.0" />
