+++ HelloRadixTests/CollibraControllerTests.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using HelloRadix.BusinessDataTypes;
+using Collibra.BusinessDataTypes;
+using Collibra.Services;
 using HelloRadix.Controllers;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
